ENH: optimize import times (#66)
It is currently very expensive to import modules that use multiple dispatch because of all of the ordering that happens at module scope.

This change creates a notion of an unresolved dispatcher which will lazily resolve the order when needed. Dispatchers begin as regular dispatchers with no functions. When `add` is called, the cache is cleared and the object moves into the unresolved state. When the `order` attribute is looked up, the order is computed and the dispatcher moves back to the resolved state. This means that exact matches (resolved in `self.funcs`) will not trigger a reorder. Reordering can be explicitly triggered by calling the `reorder` method. This same machinery applies to the `MethodDispatcher`.

This is implemented by swapping the type under the dispatcher instance to avoid a branch on every call. One simple implementation would be to have a `self._unresolved` flag on the instance; however, this would be slower than the current solution. Most multupledispatch code starts with a large amount of dispatch definitions followed by a considerably larger amount of calls to the dispatcher. I wanted to keep the per-call overhead as low as possible. This implementation also helps by batching sequences of additions and only triggers a single reorder when used.

As a small benchmark, this takes the import time of one component of our code from 7.502 seconds down to 3.210 seconds. Using the existing mutiple dispatch the top calls by total time were:
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  1742368    0.763    0.000    1.541    0.000 multipledispatch/conflict.py:7(supercedes)
   845865    0.588    0.000    1.185    0.000 multipledispatch/conflict.py:12(consistent)
      175    0.530    0.003    2.206    0.013 multipledispatch/conflict.py:24(ambiguities)
   815514    0.395    0.000    0.565    0.000 {map}
  1695752    0.343    0.000    1.847    0.000 multipledispatch/conflict.py:43(edge)
      175    0.264    0.002    2.131    0.012 multipledispatch/conflict.py:56(ordering)
```
Now, multuple dispatch is not in the top 30.

Another example is a small blaze backend implemented for a custom data format we have. This backend uses 34 calls to `dispatch`, and we spend 2.171 seconds in `restart_ordering`. The total import time for this file was 4.524 seconds. With just this change, this module's import time is down to 0.347 seconds.